Update hello world example
