+++ examples/hello_world.rs
@@ -2,18 +2,37 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-use wry::{Application, Attributes, Result};
-
-fn main() -> Result<()> {
-  let mut app = Application::new()?;
-
-  let attributes = Attributes {
-    url: Some("https://tauri.studio/".to_string()),
-    title: String::from("Hello World!"),
-    ..Default::default()
+fn main() -> wry::Result<()> {
+  use wry::{
+    webview::WebViewBuilder,
+    window::{
+      event::{Event, StartCause, WindowEvent},
+      event_loop::{ControlFlow, EventLoop},
+      window::WindowBuilder,
+    },
   };
 
-  app.add_window(attributes)?;
-  app.run();
-  Ok(())
+  let event_loop = EventLoop::new();
+  let window = WindowBuilder::new()
+    .with_title("Hello World")
+    .build(&event_loop)
+    .unwrap();
+  let _webview = WebViewBuilder::new(window)
+    .unwrap()
+    .initialize_script("menacing = 'ゴ';")
+    .load_url("https://tauri.studio")?
+    .build()?;
+
+  event_loop.run(move |event, _, control_flow| {
+    *control_flow = ControlFlow::Poll;
+
+    match event {
+      Event::NewEvents(StartCause::Init) => println!("Wry application started!"),
+      Event::WindowEvent {
+        event: WindowEvent::CloseRequested,
+        ..
+      } => *control_flow = ControlFlow::Exit,
+      _ => (),
+    }
+  });
 }
