--- examples/winit.rs
@@ -1,35 +1,0 @@
-// Copyright 2019-2021 Tauri Programme within The Commons Conservancy
-// SPDX-License-Identifier: Apache-2.0
-// SPDX-License-Identifier: MIT
-
-fn main() -> wry::Result<()> {
-  use wry::{
-    webview::WebViewBuilder,
-    window::{
-      event::{Event, StartCause, WindowEvent},
-      event_loop::{ControlFlow, EventLoop},
-      window::WindowBuilder,
-    },
-  };
-
-  let event_loop = EventLoop::new();
-  let window = WindowBuilder::new().build(&event_loop).unwrap();
-  let webview = WebViewBuilder::new(window)
-    .unwrap()
-    .initialize_script("menacing = 'ゴ';")
-    .load_url("https://tauri.studio")?
-    .build()?;
-
-  event_loop.run(move |event, _, control_flow| {
-    *control_flow = ControlFlow::Poll;
-
-    match event {
-      Event::NewEvents(StartCause::Init) => println!("Hello"),
-      Event::WindowEvent {
-        event: WindowEvent::CloseRequested,
-        ..
-      } => *control_flow = ControlFlow::Exit,
-      _ => (),
-    }
-  });
-}
